RandomArt peformance improvements
By using svg defs and "use" element
